Add PatternToNumber and NumberToPattern
PatternToNumber and NumberToPattern are preliminary functions to solve
the pg. 39 Charging Station (The frequency array) for a faster version
of FrequentWords to count the most frequent k-mer in a character string.
This commit solves the Charging Station on pg. 41 (Converting patterns
to numbers and vice-versa) which lays the groundwork for "The frequency
array".

PatternToNumber converts a nucleotide sequence into an integer giving
its lexicographic order (starting at index 0). It uses the function
SymbolToNumber and solves Coding Challenge 1L.

NumberToPattern converts an inteer into a nucleotide sequence of a given
length k with the corresponding lexicographic index (again starting at
0). It is the inverse to PatternToNumber. It uses the function
NumberToSymbol and solves Coding Challenge 1M.