Add ComputingFrequencies and FasterFrequentWords
ComputingFrequencies used PatternToNumber to generate a frequency array
of k-mers in a given text for a given value of k. FasterFrequentWords
uses the frequency array computed by ComputingFrequencies to return a
vector of the most frequent k-mers in a text for a given value of k.

ComputingFrequences solves Code Challenge 1K. This commit resolves the
Charging Station "The frequency array" on pg. 39.