Add ClumpFinding and WorseClumpFinding
ClumpFinding solves the Clump Finding Problem (Code Challenge 1E). It
finds (L,t)-clumps of k-mers in a string genome (k-mers that occur at
least t times in some window of genome of length L).

WorseClumpFinding does the same thing but is a less efficient algorithm.
These functions also resolve the Charging Station on pg. 44 "Solving the
Clump Finding Problem". Note that WorseClumpFinding corresponds to the
pseudocode for ClumpFinding on pg. 45 and ClumpFinding corresponds to
the pseudocode for BetterClumpFinding on pg. 46.