Add FindMinimumSkew
FindMinimumSkew solves the Minimum Skew Problem (Code Challenge 1F). It
finds the location of the minimum skew in a string Genome. The Skew of
the ith character/position in the string is defined as the difference between the
total number of occurrences of G and the total number of occurrences of
C in the first i nucleotides.

Finding the location of the minimum skew is suggestive of the location
of origin oriC since the lagging half-strand is low in C due to the
potential for deamination to T when single stranded whereas the leading
half-strand (of the same DNA strand, traveling in opposite directions
from oriC) is low in G to avoid deamination of C in the complementary
lagging half-strand. Thus the skew increases on the lagging half-strand
and decreases on leading half-strand. The minimum skew occurs when the
skew switches from decreasing to increasing, ie when we switch from the
leading half-strand to the lagging half-strand. This occurs at oriC.