Add HammingDistance and ApproximatePatternMatch
HammingDistance computes the Hamming distance of two strings of equal
length (ie the number of mismatches between the two strings).
HammingDistance solves the Hamming Distance Problem (Code Challenge 1G).

HammingDistance is not vectorized (can only accept single strings as
input and not character vectors). It might be worthwhile to vectorize
HammingDistance to simplify its application in ApproximatePatternMatch.
A preliminary approach might iterate through the items in the list
created by strsplit, however there is the issue of p and q as vectors
having different lengths.

ApproximatePatternMatch finds all approximate occurrences of a pattern
in a string (ie matches can have a Hamming distance of at most d from
the pattern). It solves the Approximate Pattern Matching Problem (Code
Challenge 1H).

ApproximatePatternCount was also added. It might be useful for solving
Code Challenge 1I: Frequent Words with Mismatches Problem.

To parallel PatternCount, MatchPattern was renamed to PatternMatch.
ApproximatePatternMatch and ApproximatePatternCount are the analagous
functions.