Restructure directory as an R package
Restructures the directory as an R package to allow for generation of
documentation using Roxygen2. Implementation of algorithms/exercises as
functions goes into the appropriate chapter file in the R directory,
datasets from the book and data for Rosalind exercise checks goes in the
data directory, and scripts to execute the exercise checks go in the
scripts directory.