Add Neighbors
Neighbors computes the d-neighborhood of a given string, where the
d-neighborhood of a k-mer is the set of all k-mers that are at most a
Hamming distance of d from the string. It solves Code Challenge 1N.

Neighbors uses a recursive algorithm and an alternative iterative
algorithm for generating the d-neighborhood of a string is provided by
the function IterativeNeighbors which relies on the function
ImmediateNeighbors for generating 1-neighborhoods.

Together these functions resolve the Charging Station on pg. 49
"Generating the neighborhood of a string".