Add FrequentWordsWithMismatches
FrequentWordsWithMismatches finds the most frequent k-mer with up to d
mismatches in a given character string. It solves the Frequent Words
with Mismatches Problem (Code Challenge 1I). The algorithm used is the
one outlined in the Charging Station "Solving the Frequent Words with
Mismatches Problem" on pg. 47 and therefore resolves that Charging
Station.

FindingFrequentWordsWithMismatchesBySorting was also added and it does
the same thing as FrequentWordsWithMismatches but uses an algorithm that
relies on sorting of the lexicographic indices of k-mers appearing in
the string with up to d mismatches. This algorithm is outlined in the
Charging Station "Finding frequent words with mismatches by sorting" on
pg. 51 and resolves that Charging Station.