Add FrequentWordsWithMismatchesAndComplements
The function FrequentWordsWithMismatchesAndComplements finds the most
frequent k-mers (with mismatches and reverse complements) in a string.
That is, it finds all k-mers that maximize the sum of matches with up to
d mismatches for itself and its reverse complement. It solves the
Frequent Words with Mismatches and Reverse Complements Problem (Code
Challenge 1J).