Add StringDistance and MedianString (2B)
The function StringDistance expands on HammingDistance by allowing for
the computation of the distance between strings of unequal length and
between a string and a collection of strings of equal or greater length.
It's possible to remove the restriction that the collection of strings
must be of equal or greater length by always identifying which is the
shorter string and proceeding as usual for the computation of the
distance between two strings of different lengths rather than assuming
that the lone string is the shorter string.

StringDistance is used in the implementation of MedianString since a
median string Pattern for a collection Dna is one which minimizes the
distance d(Pattern, Dna). MedianString solves the Median String Problem
(Code Challenge 2B).