Make ColumnCount produce an error on vector input
In the implementation of GreedyMotifSearch (to be completed), it is
necessary to iterate over the rows of a motif matrix, building the
matrix up iteratively. The iteration starts at a single row, but when
subsetting a 1-dimensional subset of a matrix (e.g. part of a column or
row) or a 0-dimensional subset (i.e. a single element) R reduces the
dimensionality, removing the dim attribute of the subset. To prevent
this behavior e.g. matrix[i, ] should be replaced by matrix, i, drop =
FALSE].

This single row subsetting would result in a vector being passed to
MotifProfile and thus to MotifCount.  This would in turn result in an
error when attempting to use apply() on a vector ("dim(X) must have a
positive length"). Alternatively, a bug occurred by coercing the input
to MotifProfile or MotifCount to a matrix, because as.matrix(row) or
matrix(row) actually fills in the matrix by columns as the default.
Thus, undesired results would be produced.

MotifProfile and MotifCount now both check to ensure that the input is a
matrix and produce an error if the dim attribute is missing.