Add GreedyMotifSearch (2D)
GreedyMotifSearch uses a greedy algorithm to search a collection of
strings for a motif of a specified length. It returns a matrix of the
motifs it finds in the collection with the lowest score (as calculated
by MotifScore). GreedyMotifSearch solves Code Challenge 2D and the code
in ch2_script.R was used to solve the problem on Rosalind.