Add GreedyMotifSearchPseudocounts
The function GreedyMotifSearchPseudocounts is an improvement over
GreedyMotifSearch that incorporates pseudocounts. It does this by adding
a small value to each element of the counts matrix to ensure that there
are no nonzero elements in the profile matrix (to be precise, 1 is added
to each element of the counts matrix). GreedyMotifSearchPseudocounts
completes Code Challenge 2E.

The code in ch2_script.R was used to sucessfully solve 2E on Rosalind.