Add more functions for working with motif matrices
This commit contains some additional functions for working with motif
matrices in preparation for writing RandomizedMotifSearch.

- MotifStrings is the inverse function to MotifMatrix. It takes a motif
matrix and generates a character vector corresponding to the collection
of strings represented by the motif matrix.

- Motifs is an extension of ProfileMostProbableString. Given a motif
profile and collection of strings (or the corresponding motif matrix),
Motifs returns the profile-most probable string in each string in the
collection (row in the matrix).

- MotifProfilePseudocounts is a modification of MotifProfile that uses
pseudocounts in the count matrix to ensure there are no zeros in the
profile matrix. GreedyMotifSearchPseuodocunts was modified to use
MotifProfilePseudocounts instead of its own internal implementation of
pseudocounts.