Add RandomSubstrings
The function RandomSubstrings randomly selects a substring for each
element of a character vector. It can accept a vector of probability
weights for the selectable positions for the start of the substrings.
Ideally RandomSubstrings would accept a matrix of probabilities so that
each element uses the corresponding row as for its probability weights.
However, this is not necessary for GibbsSampler (which will use
RandomSubstrings), so for now it only accepts a single vector for
probabilities which it uses for every element in the strings vector.

RandomizedMotifSearch was updated to use RandomSubstrings internally for
its randomization.