Add GibbsSampler (2G)
The function GibbsSampler searches a collection of strings for a motif
using a randomized algorithm. It returns a motif matrix containing the
corresponding motifs found in each string of the collection.
GibbsSampler solves Code Challenge 2G. The code in ch2_script.R was used
to solve the problem on Rosalind.

The documentation for ProfileRandomString was extended to include a
description of the return value which explains that it returns a string
instead of the corresponding motif matrix. An earlier version of
GibbsSampler failed to use MotifMatrix on the result of
ProfileRandomString and therefore failed to produce low scoring motifs.

The scripts for previous Code Challenges was modified to use the new
name for MotifString (instead of MotifStrings).