Add StringFromGenomePath (3B)
The function StringFromGenomePath reconstructs a string from its genome
path. A genome path is a sequence of k-mers such that the last k-1
symbols of a k-mer are equal to the first k-1 symbols of the next k-mer
in the sequence. StringFromGenomePath completes Code Challenge 3B and
the code in the script file ch3_script.R successfully solved the problem
on Rosalind.