Add OverlapGraph (3C)
The function OverlapGraph generates the overlap graph of a collection of
k-mers. The overlap graph is a directed graph where an edge exists from
one k-mer to another if the suffix of the first is equal to the prefix
of the second. The functions Suffix and Prefix are also added to check
this condition in OverlapGraph. OverlapGraph resolves Code Challenge 3C
(Overlap Graph Problem).

The code in ch3_script.R was used to successfully solve Code Challenge
3C on Rosalind.