Add PathGraph
The function PathGraph accepts as input a string and an integer k and
outputs a graph that has all the k-mers in the string as nodes and an
edge from each k-mer to the subsequent k-mer.

PathGraph is used to construct the De Bruijn graph of a string by
"gluing" together identical k-mers into the same node.