Add DeBruijnGraph (3D, 3E)
The function DeBruijnGraph constructs the De Bruijn graph for either a
single string or a collection of k-mers. In the case of a single string,
an integer k must be provided which specifies the length of substrings
to be used. The De Bruijn graph is then constructed by gluing together
the identical nodes of the corresponding path graph. This function of
DeBruijnGraph resolves Code Challenge 3D: De Bruijn Graph from a String
Problem.

If an integer k is not passed to DeBruijnGraph then the function
interprets its input as a collection of k-mers (i.e. substrings of
length k of some string) and uses this collection to construct the De
Bruijn graph. The prefixes and suffixes of the k-mers are used as nodes
while the k-mers label the edges from their prefixes to their suffixes.
This function of DeBruijnGraph resolves Code Challenge 3E: De Bruijn
Graph from k-mers Problem.

The code used to solve these challenges on Rosalind is included in
ch3_script.R.