Add EulerianCycle and accessory functions (3F)
EulerianCycle finds an Eulerian cycle in a balanced, strongly connected
graph using the algorithm set forth in the original proof of Euler's
Theorem. It resolves Code Challenge 3F.

This commit also adds the accessory functions AdjacencyListToGraph and
PrintPath which are used to interface with Rosalind to solve the Code
Challenge. AdjacencyListToGraph is used to convert the text adjacency
lists from Rosalind into graphs as list structures which can be fed into
EulerianCycle. PrintPath converts the character matrix output of
EulerianCycle into a single string in the format necessary for Rosalind.

This commit also add the function ConnectedNodes for finding nodes
adjacent to a particular node, but it appears this function was never
used for anything.

The code used to solve problem 3F on Rosalind is also included.