Add EulerianPath (3G)
The function EulerianPath finds an Eulerian path in a nearly balanced
graph. An Eulerian path is a path that traverses each edge of a graph
exactly once. EulerianPath resolves Code Challenge 3G and the code used
to solve this challenge on Rosalind is included in ch3_script.R.

The function GraphDegree is also added, which is used to find any
unbalanced nodes of a graph. This is used to identify which nodes need
an edge between them in a nearly balanced graph so that the graph can be
come balanced and strongly connected. EulerianCycle can then be used to
find an Eulerian cycle, which is modified to produce an Eulerian path by
removing the previously added edge.

GraphDegree saves the degree of a graph in a tibble. A dataframe type
object is preferred since the names of the nodes (the first column of
the tibble) should be a character vector while the indegree and
outdegree (the second and third columns) should be integers. However
using a tibble requires loading tidyverse thus it may be better to
simply use a dataframe or character matrix since indegree and outdegree
can be coerced to integers as necessary.