Make EulerianPath work with de Bruijn graphs
GraphDegree and EulerianPath implicitly assumed that graph$edges of its
input consisted only of the columns "node1" and "node2". For determining
the nodes, GraphDegree found all the unique elements of graph$edges, but
this might contain edge labels in the case of a de Bruijn graph
generated by DeBruijnGraph.

EulerianPath assumed that there were only two columns in graph$edges and
hence added the balancing edge between the unbalanced nodes by using
rbind and the vector consisting of only the two edges. If there are more
than two columns in graph$edges this results in repeating the elements
of the two edge vectors and does not guarantee that the nodes are place
in the correct column. Instead a row of NAs is added and the node names
are placed in the appropriate location to fill in the edge.

This change is necesary to reconstruct a string from its k-mer
composition in StringFromComposition (3H).