Allow EulerianPath to also find Eulerian cycles
Previously EulerianPath check to make sure that the input graph has
exactly 2 unbalanced nodes and that the graph is therefore nearly
balanced. However, EulerianPath should generalize EulerianCycle since a
cycle is also a path. Thus EulerianPath now checks whether the graph is
balanced or nearly balanced and produces an error if neither case
applies. If the graph is balanced then an Eulerian cycle is found.