Fix StringFromComposition for circular strings
An optional logical parameter 'circular' was added to
StringFromComposition which allows specification of whether the string
to be reconstructed is circular. If the string is circular then the
output string needs to be adjusted by trimming the end of the string
that overlaps with the beginning. This change will allow
StringFromComposition to work for finding k-universal circular binary
strings (Code Challenge 1I).