Allow functions to use amino acid mass vectors
ConvolutionCyclopeptideSequencing constructs an alphabet for use in
sequencing a cyclopeptide from a noisy experimental spectrum. This
alphabet allows for amino acids of arbitrary mass between 57 and 200,
which requires the use of numbers to represent amino acids by their
masses instead of the single letter codes. Thus, collections of peptides
would be numeric vectors contained in lists. This commit updates the
peptide sequencing functions to treat peptides in this manner.